requirements.txt
# Sign Language Recognition - All Dependencies
# Includes both training and backend requirements

# ===== Training Dependencies =====
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tf2onnx>=1.16.0
onnx>=1.14.0
Pillow>=10.0.0
mediapipe>=0.10.0
tqdm>=4.66.0
scipy>=1.10.0

# ===== Backend API Dependencies =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
onnxruntime>=1.16.0
python-multipart>=0.0.6

# ===== Optional: GPU Support =====
# Uncomment for GPU-accelerated inference:
# onnxruntime-gpu>=1.16.0

# ===== Testing Dependencies =====
requests>=2.31.0
